+++ fizz-core/src/main/java/we/stats/ResourceStat.java
@@ -20,6 +20,7 @@ package we.stats;
 import java.math.BigDecimal;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
@@ -50,7 +51,8 @@ public class ResourceStat {
 	/**
 	 * Concurrent requests
 	 */
-	private AtomicLong concurrentRequests = new AtomicLong(0);
+//	private AtomicLong concurrentRequests = new AtomicLong(0);
+	private AtomicInteger concurrentRequests = new AtomicInteger(0);
 
 	private ReentrantReadWriteLock rwl1 = new ReentrantReadWriteLock();
 	private ReentrantReadWriteLock rwl2 = new ReentrantReadWriteLock();
@@ -95,16 +97,16 @@ public class ResourceStat {
 		try {
 			boolean isExceeded = false;
 			if (maxCon != null && maxCon.intValue() > 0) {
-				long n = this.concurrentRequests.get();
+				int n = this.concurrentRequests.get();
 				if (n >= maxCon.longValue()) {
 					isExceeded = true;
 					this.incrBlockRequestToTimeSlot(timeSlotId);
 				} else {
-					long conns = this.concurrentRequests.incrementAndGet();
+					int conns = this.concurrentRequests.incrementAndGet();
 					this.getTimeSlot(timeSlotId).updatePeakConcurrentReqeusts(conns);
 				}
 			} else {
-				long conns = this.concurrentRequests.incrementAndGet();
+				int conns = this.concurrentRequests.incrementAndGet();
 				this.getTimeSlot(timeSlotId).updatePeakConcurrentReqeusts(conns);
 			}
 			return !isExceeded;
@@ -118,7 +120,7 @@ public class ResourceStat {
 	 * 
 	 */
 	public void decrConcurrentRequest(long timeSlotId) {
-		long conns = this.concurrentRequests.decrementAndGet();
+		int conns = this.concurrentRequests.decrementAndGet();
 		this.getTimeSlot(timeSlotId).updatePeakConcurrentReqeusts(conns);
 	}
 
@@ -335,11 +337,11 @@ public class ResourceStat {
 		this.timeSlots = timeSlots;
 	}
 
-	public AtomicLong getConcurrentRequests() {
+	public AtomicInteger getConcurrentRequests() {
 		return concurrentRequests;
 	}
 
-	public void setConcurrentRequests(AtomicLong concurrentRequests) {
+	public void setConcurrentRequests(AtomicInteger concurrentRequests) {
 		this.concurrentRequests = concurrentRequests;
 	}
 }
